[UPDATE] fix #855

+++ src/main/scala/com/github/unchama/seichiassist/util/exp/IExperienceManager.java
@@ -5,23 +5,23 @@ import org.bukkit.entity.Player;
 public interface IExperienceManager {
     Player getPlayer();
 
-    void changeExp(int amt);
+    void changeExp(long amt);
 
     void changeExp(double amt);
 
-    void setExp(int amt);
+    void setExp(long amt);
 
     void setExp(double amt);
 
-    int getCurrentExp();
+    long getCurrentExp();
 
-    boolean hasExp(int amt);
+    boolean hasExp(long amt);
 
     boolean hasExp(double amt);
 
-    int getLevelForExp(int exp);
+    long getLevelForExp(long exp);
 
-    int getXpNeededToLevelUp(int level);
+    long getXpNeededToLevelUp(int level);
 
-    int getXpForLevel(int level);
+    long getXpForLevel(int level);
 }
